# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server EHRS server.",
    "version": "1.0.0",
    "title": "EHRS",
    "contact": {
      "email": "Admin@ehrs.ir"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/generics/import": {
      "post": {
        "tags": [
          "Generic"
        ],
        "summary": "import generic items from xlsx file",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "generic",
            "in": "formData",
            "description": "Updated file of the generic",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          },
          "401": {
            "description": "product validation failed"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
}